requirements.txt
fastapi
uvicorn
pandas
openpyxl
jinja2
selenium
webdriver-manager
python-dotenv
requests